refactor(#826): split force_release.rs into mod.rs + gc.rs (C4)
scope follows dispatch spec t-20260515164008976728-0

CI at C3 (018a071) failed across all 3 platforms at the
`mcp_handler_files_under_500_loc` integration test (in
`tests/file_size_invariant.rs`): src/mcp/handlers/force_release.rs
grew to 1166 LOC, exceeding the 750-LOC handler-file ceiling.

Pure code-motion refactor (no semantic change). The #826 GC code
moves to a new sibling module:

- `src/mcp/handlers/force_release/mod.rs` — 579 LOC. Keeps the
  pre-#826 handler surface: `handle_force_release_worktree`,
  `RebaseCleanOutcome`, `rebase_clean_self`, and the 12 pre-existing
  tests + `tmp_home` / `seed_daemon_worktree` fixtures.
- `src/mcp/handlers/force_release/gc.rs` — 626 LOC. Holds the L2 GC
  helpers (`GcOutcome`, `prune_git_metadata_for_agent`,
  `list_worktrees_bypass_shim`, `discover_source_repo_candidates`,
  `collect_source_repos_from_worktree_tree`, `paths_match`,
  `canonicalize_via_parent`) + the 5 #826 tests + the
  `seed_disbanded_agent_with_git_metadata` fixture.

mod.rs declares `mod gc;` and calls `gc::prune_git_metadata_for_agent`.
Visibility: helpers now `pub(super)` so mod.rs can reach them; the
helper module itself is private to `force_release`.

Verified locally:
- `cargo test --test file_size_invariant` — GREEN (was the failing
  step on CI). mod.rs 579 + gc.rs 626 both under 750.
- `cargo test --bin agend-terminal force_release_worktree` — 17/17
  passing (5 gc::tests + 12 mod::tests).
- `cargo test --tests --features tray` — all integration test
  binaries green.
- `cargo fmt --check` + `cargo clippy --all-targets --features tray
  -- -D warnings` clean.

Pre-push checklist updated per lead's dispatch (add `cargo test
--tests --features tray` to catch invariant integration tests; the
prior `cargo test --bin agend-terminal` ran binary unit tests only).


Agend-Agent: fixup-dev
Agend-Task: t-20260515164008976728-0
Agend-Branch: fix/826-force-release-worktree-gc-disbanded
Agend-Issued-At: 2026-05-15T17:00:05.747481+00:00
